# Server&Client/server/expServer/docs/swagger.json
{ "openapi": "3.0.0",
    "info": {
      "title": "CAB230 : Web Computing API",
      "description": "This API is to be used during development of assignments for the QUT CAB230 Web Computing Unit. This instance is locally hosted",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "https://localhost"
      }
    ],
    "paths": {
      "/register": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Registers a new user account",
          "description": "This is your starting point.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "required": true,
                      "example": "demo-user@gmail.com"
                    },
                    "password": {
                      "type": "string",
                      "required": true,
                      "example": "demouserpassword"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Created",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "yay! you've successfully registered your user account :)"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "oops! It looks like that user already exists :("
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "Login with an existing user account",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "required": true,
                      "example": "demo-user@gmail.com"
                    },
                    "password": {
                      "type": "string",
                      "required": true,
                      "example": "demouserpassword"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "access_token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjo1Nzg2LCJlbWFpbCI6ImRlbW8xMjM0QGdtYWlsLmNvbSJ9LCJpYXQiOjE1NTUzMjQ2MDIsImV4cCI6MTU1NTQxMTAwMn0.mo85iKuAQ_h8n8-6pkeNNs5CCnV0G8hScxhi_ml_Q6c"
                      },
                      "token_type": {
                        "type": "string",
                        "example": "Bearer"
                      },
                      "expires_in": {
                        "type": "intager",
                        "example": 86400
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "invalid login - bad password"
                      }
                    }
                  }
                }
              }
            }
          },
          "tags": [
            "Authentication"
          ]
        }
      },
...
    },
...
  }